requirements.txt
fastapi
uvicorn
python-multipart
openai
pyyaml
